Move indentation-aware classes to separate module
